requirements.txt
# Core Data & ML
nfl-data-py==0.3.3
pandas==1.5.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# Streaming Infrastructure
kafka-python==2.0.2
pyspark==3.5.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
